Massive refactor but its cool
