+++ readme.md
@@ -1,5 +1,2 @@
 # Chattr is a Discord bot that can access ChatGPT
-
-# NOTE: This is not the proper way to make a discord bot
-
-# No Moderation Privileges: https://discord.com/api/oauth2/authorize?client_id=1085319165650411570&permissions=3147776&scope=bot
+# No Moderation Privileges Invite Link: https://discord.com/api/oauth2/authorize?client_id=1085319165650411570&permissions=3147776&scope=bot
