+++ res/configs.json.json
@@ -1,0 +1,46 @@
+{
+    "474096541142089728": {
+        "conversee": null,
+        "last_sender": null,
+        "current_status": "Off",
+        "model": "gpt-3.5-turbo",
+        "in_voice": false,
+        "voice_client": null,
+        "tts": false,
+        "tts_upgrade": false,
+        "moderation": false
+    },
+    "1065302721479049368": {
+        "conversee": null,
+        "last_sender": null,
+        "current_status": "Off",
+        "model": "gpt-3.5-turbo",
+        "in_voice": false,
+        "voice_client": null,
+        "tts": false,
+        "tts_upgrade": false,
+        "moderation": false
+    },
+    "1083899422770995321": {
+        "conversee": null,
+        "last_sender": null,
+        "current_status": "Off",
+        "model": "gpt-3.5-turbo",
+        "in_voice": false,
+        "voice_client": null,
+        "tts": false,
+        "tts_upgrade": false,
+        "moderation": false
+    },
+    "1085772115736797205": {
+        "conversee": null,
+        "last_sender": null,
+        "current_status": "Off",
+        "model": "gpt-3.5-turbo",
+        "in_voice": false,
+        "voice_client": null,
+        "tts": false,
+        "tts_upgrade": false,
+        "moderation": false
+    }
+}
