+++ start_server.sh
@@ -1,0 +1,1 @@
+python3 bin/bot.py
